refactor: 로또 결과에 대한 state를 파생상태로 변경
